errors_corrected

+++ pipelines/inference_pipeline.py
@@ -132,13 +132,13 @@ def main():
         features     = [
             Feature("pickup_location_id", "string"),
             Feature("pickup_hour",        "timestamp"),
-            Feature("predicted_rides",    "int"),
+            Feature("predicted_rides",    "int64"),
         ],
     )
 
     # ── 1️⃣2️⃣  Cast to the FG schema and insert
     preds["pickup_location_id"] = preds["pickup_location_id"].astype(str)
-    preds["predicted_rides"]    = preds["predicted_rides"].astype("int32")
+    preds["predicted_rides"]    = preds["predicted_rides"].astype("int64")
 
     pred_fg.insert(preds, write_options={"wait_for_job": False})
     print("✅ Inference complete — predictions up to", preds["pickup_hour"].iloc[0])
